+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/MetroCompilerTest.kt
@@ -379,12 +379,6 @@ abstract class MetroCompilerTest {
                   this@toPluginOptions.generateStaticAnnotations,
                 )
               }
-              BINDING_CONTRIBUTIONS_AS_CONTAINERS -> {
-                processor.option(
-                  entry.raw.cliOption,
-                  this@toPluginOptions.bindingContributionsAsContainers,
-                )
-              }
               MEMBER_NAMING_STRATEGY -> {
                 processor.option(entry.raw.cliOption, this@toPluginOptions.memberNamingStrategy)
               }